chore(docker): Add bind mounts for pyproject.toml and poetry.lock
Added bind mounts for pyproject.toml and poetry.lock to enable real-time
updates during development. Ensure that the permissions for these files
are set to 666 on the host machine to avoid access issues inside the
container, as the container runs with a non-root user that requires read
and write access.